// src/constants/enums.ts
export enum EventStatuses {
  UPCOMING = "upcoming",
  ONGOING = "ongoing",
  COMPLETED = "completed",
}

export enum RoundStatuses {
  UPCOMING = "upcoming",
  ONGOING = "ongoing",
  SHOW_RESULT = "show_result",
  COMPLETED = "completed",
}

export enum UserRoles {
  LEAD_PLAYER = "leadPlayer",
  COMMON_PLAYER = "player",
}

export enum TextMessages {
  LIKE = "likes the drawing",
  DISLIKE = "dislikes the drawing",
}

export enum ConnectionState {
  CONNECTED = "connected",
  CONNECTING = "connecting",
  DISCONNECTED = "disconnected",
  RECONNECTING = "reconnecting",
}